patch broken file selector; add XRI samples
the real issue with the file selector is that the Slider FilesSlider
(in line 257) doesn't exist in the game hierarchy. where did it go?
is it okay to just load user files if it doesn't exist?

can't do anything without the files, so though idk if this is the right
way to patch, it works for now